+++ schema/mutations/deleteQuestionnaire.js
@@ -1,0 +1,19 @@
+const models = require("../../models");
+const { Questionnaire } = require("../types");
+const { GraphQLID } = require("graphql");
+
+module.exports = {
+  type : Questionnaire,
+
+  args : {
+    id : {
+      type : GraphQLID
+    }
+  },
+
+  resolve(_, { id }) {
+    return models.Questionnaire
+      .findById(id)
+      .then(questionnaire => questionnaire.destroy().then(() => questionnaire));
+  }
+}
