+++ schema/types/Answer.js
@@ -1,0 +1,9 @@
+const models = require("../../models");
+const { GraphQLObjectType } = require("graphql");
+const { attributeFields } = require("graphql-sequelize");
+
+module.exports = new GraphQLObjectType({
+  name : "Answer",
+  description : "An answer",
+  fields : attributeFields(models.Answer)
+});
