+++ schema/types/index.js
@@ -1,0 +1,6 @@
+module.exports = {
+  Answer : require("./Answer"),
+  Question : require("./Question"),
+  Page : require("./Page"),
+  Questionnaire : require("./Questionnaire")
+};
